doc-gen-api/pyproject.toml
[project]
name = "doc-gen-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "openai>=2.7.2",
    "pypandoc>=1.16",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.4",
    "sse-starlette>=3.0.3",
]

[tool.ruff]
indent-width = 4
exclude = ["prompts/*.py"]


[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
extend-select = ["E501"]
